requirements.txt
Flask>=2.0
kafka-python>=2.0
